+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+- 1.0.3
+  - Bug fix
+    - Handling of possible null Activity from ReactApplicationContext
+    - Additional null checks on Consent Manager methods
+  - Sample App
+    - Gradle Android versions updated to latest
 - 1.0.2
   - Bug fix
     - Fixed public trackEvent to call trackEvent on Tealium instance